Changes to tx construction
With these changes, we build the transaction in place directly, and poke
in the signing data, rather than repeatedly re-building the tx from
its inputs and outputs. This is both faster and simpler.

We can be better still in the future by using static dummy data for the
tx witness and input scripts when adding the initial UTXOs, as noted in
the comments.

Also added some misc FIXMEs for other noted issues and updated wally.