Update priority strings
The settings values returned by gdk for required_num_blocks are now always
one of [3, 12, 24] blocks.

When the send screen is entered, for a new tx the buttom corresponding
to the users default setting should be auto selected and the rate taken
from that block number in the current gdk fee estimates. For a bump, the
button corresponding to the next highest fee rate from the txs current
fee rate should be selected, or custom if the fee rate is >= the high
rate.